OSIS-3603 Reorganize views folder on program_management app + arrange test
