+++ program_management/urls.py
@@ -26,13 +26,12 @@
 from django.conf.urls import url
 from django.urls import include
 
-from program_management.views.excel import get_learning_unit_prerequisites_excel, \
-    get_learning_units_is_prerequisite_for_excel
 from program_management.views import groupelementyear_create, groupelementyear_delete, groupelementyear_update, \
-    groupelementyear_read, prerequisite_update, prerequisite_read, element_utilization
+    groupelementyear_read, prerequisite_update, prerequisite_read, element_utilization, groupelementyear_postpone, \
+    groupelementyear_management, excel
 
 urlpatterns = [
-    url(r'^management/$', groupelementyear_update.management, name='education_groups_management'),
+    url(r'^management/$', groupelementyear_management.management, name='education_groups_management'),
     url(r'^(?P<root_id>[0-9]+)/(?P<education_group_year_id>[0-9]+)/', include([
         url(r'^content/', include([
             url(u'^attach/', groupelementyear_create.AttachTypeDialogView.as_view(),
@@ -50,7 +49,7 @@ urlpatterns = [
         ])),
         url(r'^group_content/', groupelementyear_read.ReadEducationGroupTypeView.as_view(), name="group_content"),
         url(r'^pdf_content/(?P<language>[a-z\-]+)', groupelementyear_read.pdf_content, name="pdf_content"),
-        url(r'^postpone/', groupelementyear_update.PostponeGroupElementYearView.as_view(),
+        url(r'^postpone/', groupelementyear_postpone.PostponeGroupElementYearView.as_view(),
             name="postpone_education_group"),
     ])),
     url(r'^(?P<root_id>[0-9]+)/(?P<learning_unit_year_id>[0-9]+)/learning_unit/', include([
@@ -66,12 +65,12 @@ urlpatterns = [
     ])),
     url(
         r'reporting/(?P<education_group_year_pk>[0-9]+)/prerequisites/$',
-        get_learning_unit_prerequisites_excel,
+        excel.get_learning_unit_prerequisites_excel,
         name="education_group_learning_units_prerequisites"
     ),
     url(
         r'reporting/(?P<education_group_year_pk>[0-9]+)/is_prerequisite_of/$',
-        get_learning_units_is_prerequisite_for_excel,
+        excel.get_learning_units_is_prerequisite_for_excel,
         name="education_group_learning_units_is_prerequisite_for"
     )
 ]
